+++ src/main/java/ru/job4j/ood/lsp/parking/Car.java
@@ -2,7 +2,7 @@ package ru.job4j.ood.lsp.parking;
 
 public class Car extends Auto  {
 
-    public Car(double spaceAmount, String name) {
+    public Car(int spaceAmount, String name) {
         super(spaceAmount, name);
     }
 }
