+++ src/main/java/ru/job4j/ood/lsp/parking/Truck.java
@@ -2,7 +2,7 @@ package ru.job4j.ood.lsp.parking;
 
 public class Truck extends Auto  {
 
-    public Truck(double spaceAmount, String name) {
+    public Truck(int spaceAmount, String name) {
         super(spaceAmount, name);
     }
 }
